Init STM32 Project

+++ LAB5/Core/Inc/uart_communiation_fsm.h
@@ -16,12 +16,7 @@
 extern uint32_t last_time;
 extern UART_HandleTypeDef huart2;
 extern uint32_t ADC_value;
-extern uint32_t waiting;
-//typedef enum{
-//	IDLE,
-//	SEND_ADC,
-//	WAIT_ACK
-//} UART_State;
+extern uint8_t waiting;
 void uart_communiation_fsm();
 
 #endif /* INC_UART_COMMUNIATION_FSM_H_ */
